+++ src/main/java/de/uni_stuttgart/it_rex/course/service/mapper/written/ContentProgressTrackerMapper.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.service.mapper.written;
 
-import de.uni_stuttgart.it_rex.course.domain.written_entities.ContentProgressTracker;
+import de.uni_stuttgart.it_rex.course.domain.written.ContentProgressTracker;
 import de.uni_stuttgart.it_rex.course.service.dto.written_dtos.ContentProgressTrackerDTO;
 import org.mapstruct.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,9 +22,7 @@ public abstract class ContentProgressTrackerMapper {
      * @return the dto
      */
     public ContentProgressTrackerDTO toDTO(final ContentProgressTracker tracker) {
-        final ContentProgressTrackerDTO trackerDTO = setBasicProperties(tracker);
-
-        return trackerDTO;
+        return setBasicProperties(tracker);
     }
 
     /**
