+++ src/main/java/de/uni_stuttgart/it_rex/course/service/written/ProgressTrackingService.java
@@ -1,9 +1,8 @@
 package de.uni_stuttgart.it_rex.course.service.written;
 
 import de.uni_stuttgart.it_rex.course.domain.written.ContentReference;
-import de.uni_stuttgart.it_rex.course.domain.written_entities.ContentProgressTracker;
-import de.uni_stuttgart.it_rex.course.domain.written.ContentReference;
-import de.uni_stuttgart.it_rex.course.domain.written_entities.CourseProgressTracker;
+import de.uni_stuttgart.it_rex.course.domain.written.ContentProgressTracker;
+import de.uni_stuttgart.it_rex.course.domain.written.CourseProgressTracker;
 import de.uni_stuttgart.it_rex.course.repository.written.ContentProgressTrackerRepository;
 import de.uni_stuttgart.it_rex.course.repository.written.ContentReferenceRepository;
 import de.uni_stuttgart.it_rex.course.repository.written.CourseProgressTrackerRepository;
