+++ src/test/java/de/uni_stuttgart/it_rex/course/web/rest/written/ProgressResourceIT.java
@@ -3,8 +3,9 @@ package de.uni_stuttgart.it_rex.course.web.rest.written;
 import de.uni_stuttgart.it_rex.course.CourseServiceApp;
 import de.uni_stuttgart.it_rex.course.config.TestSecurityConfiguration;
 import de.uni_stuttgart.it_rex.course.domain.enumeration.ContentProgressState;
+import de.uni_stuttgart.it_rex.course.domain.written.ContentProgressTracker;
 import de.uni_stuttgart.it_rex.course.domain.written.ContentReference;
-import de.uni_stuttgart.it_rex.course.domain.written_entities.*;
+import de.uni_stuttgart.it_rex.course.domain.written.CourseProgressTracker;
 import de.uni_stuttgart.it_rex.course.repository.written.ContentProgressTrackerRepository;
 import de.uni_stuttgart.it_rex.course.repository.written.ContentReferenceRepository;
 import de.uni_stuttgart.it_rex.course.repository.written.CourseProgressTrackerRepository;
